requirements.txt
torch
torchvision
flask
pillow
huggingface_hub
numpy
matplotlib
